update plugins to use new detector info interface
